Optimize behaviours and add cohesion behaviour
- Change to starting values and sliders
Messed around with some starting values to make behaviours feel a bit
better.
Messed around with slider ranges as well for the same reason.

- Behaviour optimization
Added a check to start oof all behaviours to not do calculations and just
return a 0 vector as thats what all the calculations would equate to
anyway.

- Visualization Changes
Commented out seperation and cohesion radius circles, they we're cluttering the screen too much.

- Increased Agents
Increased number of agents to 500.